+++ deepmatch-plugin/src/integrationTest/java/com/aouledissa/deepmatch/gradle/integration/GenerateDeeplinkReportTaskIntegrationTest.kt
@@ -41,7 +41,7 @@ class GenerateDeeplinkReportTaskIntegrationTest {
 
     @Test
     fun generateDeeplinkReportSucceedsAndProducesHtmlReport() {
-        fixture.withAndroidProject(
+        fixture.withWorkingAndroidProject(
             buildScriptExtra = """
                 deepMatch {
                     report {